ukvm: Split block and net interface implementations
This simplifies re-use of UKVM kernel code for other upcoming platforms
which may provide different network and block implementations.
