+++ kernel/ukvm/block.c
@@ -20,45 +20,6 @@
 
 #include "kernel.h"
 
-/* ukvm net interface */
-int solo5_net_write_sync(uint8_t *data, int n)
-{
-    volatile struct ukvm_netwrite wr;
-
-    wr.data = data;
-    wr.len = n;
-    wr.ret = 0;
-
-    ukvm_do_hypercall(UKVM_HYPERCALL_NETWRITE, &wr);
-
-    return wr.ret;
-}
-
-int solo5_net_read_sync(uint8_t *data, int *n)
-{
-    volatile struct ukvm_netread rd;
-
-    rd.data = data;
-    rd.len = *n;
-    rd.ret = 0;
-
-    ukvm_do_hypercall(UKVM_HYPERCALL_NETREAD, &rd);
-
-    *n = rd.len;
-    return rd.ret;
-}
-
-static char mac_str[18];
-char *solo5_net_mac_str(void)
-{
-    volatile struct ukvm_netinfo info;
-
-    ukvm_do_hypercall(UKVM_HYPERCALL_NETINFO, &info);
-
-    memcpy(mac_str, (void *)&info, 18);
-    return mac_str;
-}
-
 /* ukvm block interface */
 int solo5_blk_write_sync(uint64_t sec, uint8_t *data, int n)
 {
